refactored unit and integration test for user (common logic separated in utills.js file)
